Update pem.js, add error checking for missing certificate
If you call the extract certificate info function without a certificate, you'll get a very crypting error message telling you about missing `--TMPFILE--`. This is because the handling of undefined files in the `spawnWrapper` leads to the file not being written to disk and thus OpenSSH being called with a missing file.